feat: detach scheduler from lean builds
