+++ crates/goose/src/acp/server_factory.rs
@@ -2,11 +2,14 @@ use crate::acp::server::{
     AcpBuiltinSelection, AcpProviderFactory, ActiveRunRegistry, GooseAcpAgent, GooseAcpAgentOptions,
 };
 use crate::agents::GoosePlatform;
+#[cfg(feature = "scheduler")]
 use crate::scheduler_trait::SchedulerTrait;
+#[cfg(feature = "scheduler")]
 use crate::session::SessionManager;
 use crate::source_roots::SourceRoot;
 use anyhow::Result;
 use std::sync::Arc;
+#[cfg(feature = "scheduler")]
 use tokio::sync::OnceCell;
 use tracing::info;
 
@@ -25,6 +28,7 @@ pub struct AcpServerFactoryConfig {
 
 pub struct AcpServer {
     config: AcpServerFactoryConfig,
+    #[cfg(feature = "scheduler")]
     scheduler: OnceCell<Arc<dyn SchedulerTrait>>,
     active_prompt_runs: ActiveRunRegistry,
 }
@@ -33,11 +37,13 @@ impl AcpServer {
     pub fn new(config: AcpServerFactoryConfig) -> Self {
         Self {
             config,
+            #[cfg(feature = "scheduler")]
             scheduler: OnceCell::new(),
             active_prompt_runs: ActiveRunRegistry::default(),
         }
     }
 
+    #[cfg(feature = "scheduler")]
     /// Start the scheduler now instead of on first client connect, so a
     /// headless `goose serve` runs scheduled jobs; on failure `create_agent`
     /// retries. No-op when the scheduler is disabled.
@@ -45,6 +51,7 @@ impl AcpServer {
         self.scheduler().await.map(|_| ())
     }
 
+    #[cfg(feature = "scheduler")]
     async fn scheduler(&self) -> Result<Option<Arc<dyn SchedulerTrait>>> {
         if !self.config.enable_scheduler {
             return Ok(None);
@@ -84,9 +91,10 @@ impl AcpServer {
     ) -> Result<Arc<GooseAcpAgent>> {
         let config = crate::config::Config::global();
         let disable_session_naming = config.get_goose_disable_session_naming().unwrap_or(false);
+        #[cfg(feature = "scheduler")]
         let scheduler = self.scheduler().await?;
+        #[cfg(feature = "scheduler")]
         if let Some(scheduler) = &scheduler {
-            // Listing syncs from storage, registering jobs persisted by other processes.
             scheduler.list_scheduled_jobs().await;
         }
 
@@ -122,7 +130,10 @@ impl AcpServer {
             goose_platform: self.config.goose_platform.clone(),
             additional_source_roots: self.config.additional_source_roots.clone(),
             session_cwd,
+            #[cfg(feature = "scheduler")]
             scheduler,
+            #[cfg(not(feature = "scheduler"))]
+            scheduler: None,
             active_prompt_runs: self.active_prompt_runs.clone(),
         })
         .await?;
