+++ crates/goose/src/scheduler/mod.rs
@@ -1,0 +1,3 @@
+mod full;
+
+pub use full::*;
